Nuke SCSS variable $font-color-primary
Use $font-color instead. It were the same colors in the dark theme
anyway. In the light theme it produced colors that were slightly
off. E.g. in the overview of rulesets the texts were black and the
numbers grey.

Also rename $font-color-primary-hover to $font-color-hover and resolve a
TODO for message boxes (used e.g. by the global setting debug livestatus
queries).

CMK-8194

Change-Id: I758c9d4d6998d52242fb66a56a5842ac73a49af6