Fix conditionals in escape_attribute
Since type(value) was used in isinstance checks they always evaluated to
False. I.e. for the common case str all conditionals were always
evaluated. Now, isinstance is used on the value and the str and HTML
cases are moved to the top. In tests this resulted in a 25 - 30%
performance improvement.

CMK-8035

Change-Id: I7adc9577f2ecdcbb2d3854c179ac6675d3dcbee7