11814 FIX smart check: Evaluate command timeouts as counter
Previously, the metric "Command_Timeout" of the smart check was evaluated in the following manner:
If the counter deviated from the discovered value, the check went {CRIT}.

This led to a lot of false posivitves, as the counter may also increase after a simple reboot, which is no faulty situation.
The rate however gives a better evaluation of the state.
As from this werk, we allow all rates below 100 Command Timeouts per hour.

SUP-7205

Change-Id: I06f98b19b023139869e0f960216f8cd7f79606f7