Python3: Package was removing whole build dir
The target was executing `rm -f -r $(PACKAGE_BUILD_DIR)/` instead
of removing the intentioned `$(PACKAGE_BUILD_DIR)/$(PYTHON3_DIR)`.

This is a quick fix for the specific problem. We should better rename
the OMD package to use the default unpack target implementation from
omd/packages/packages.make. This is currently not possible, because the
OMD package name does not match the source archive file.

Change-Id: I6f81ab4ecce0919bfdf37db7cbd420f6d4c8eab3