12996 FIX Performance graphs always use default color styles
Before checkmk 2.0.0 it was possible for users to configure the background,
foreground and canvas color of the performance graphs. This was ralery useful
and with the posiblility to switch between theme having a fixed set of style
custom set colors was unpleasing on theme changes.

Although the option to set these colors was removed the web renderer could still
read them out of the config. This was particularly anoying on modified service
views, which contain performance graphs and which had saved their default pre
2.0 color default, which don't match the new color themes in version 2.0.0

This Werk enforces checkmk's color themes for all Performance graphs and
discards any previous custom setup.

Enforce in defaults. Remove treatment on JS render, as they are
transparent now.

SUP-6084

Change-Id: I10cfc5ee9da7f1698d7a5977caefec466dd86010