Fix conditionals in escape_attribute
Since type(value) was used in isinstance checks they always evaluated to
False. I.e. for the common case str all conditionals were always
evaluated. Now, isinstance is used on the value and the str and HTML
cases are moved to the top. In tests this resulted in a 25 - 30%
performance improvement.

Furthermore, the case for bytes is removed and one for float and
LazyString are added. If no conditional matches an exception is thrown.
Any call sites that raise an exception in the future should be fixed.

CMK-8035

Change-Id: I60e6f87caad1a0f4061f1ed2efb8b9e1205b53aa