13146 FIX chrony: warn if synchronization is lost
The plugin now detects if the synchronization to the reference source
is lost. Before, the plugin returned OK in this case. Now, WARN/CRIT is
returned if synchronization is lost for a longer timeperiod. The
default values are 30 and 60 minutes and can be configured with the
option "Phases without synchronization" in the ruleset "State of NTP
time synchronization". Note that "NTP time" services are affected as
well by this ruleset.

SUP-7320

Change-Id: I3191732b7ef2faef76ce54fc4e086c2242c71c1b