Fix SNMP integration test on debian 10
For some reason the SNMP OID resolving of the `netsnmp` Python module is
working a little different on Debian 10 than on other distros. The
previously used name `sysDescr.0` could not be resolved since the latest
changes to the net-snmp packaging logic from commit
(8cacb48c5703f5b803c0e2eaa4a165d3b6d66e9e).

However, the symbolic name resolving of OIDs is working. All the
relevant MIBS shipped with Checkmk can be found. But the resolving of
the symbolic names now only works when you use the full name (including
the `MIB::` prefix).

Since Checkmk is not relying on SNMP OID translation it seems be OK to
not track down this issue and just change the OID to be queried in the
snmpsimd fixture to the symbolic name `SNMPv2-MIB::sysDescr.0`.

Should it turn out that there are more problems with the change
mentioned above, please revert these changes immediately without further
analyzes.

Change-Id: Id59d4b7390482d6068d6b07bddea974511298b1e