13075 FIX agent_prometheus: Remove web port option from custom url option
In the previous version, the Prometheus Web Port option was only
relevant for the host name and ip address connection options. This
led to confusion as it was possible to specify the Web Port but
did not apply to the custom url connection option. This werk makes
the port option more specific.

Change-Id: Ic140fd5dee552045be6e5480c8bf885d2edcadea