11815 Bakery plugin for Windows Oracle Monitoring
Previously, only the unix-like operating systems did have a bakery plugin for baking a configuration for mk_oracle.
Now the plugin also supports baking the configuration for the windows plugin (mk_oracle.ps1).
However, please note that not all features of the bash plugin (mk_oracle) are available for windows yet.
This is noted in the rulespec "ORACLE Databases (Linux, Solaris, AIX, Windows)".

Change-Id: Ic0260dbaa3256c611859530734c0a1e0d2a8217b