hooks vanished after unregister_plugin_hooks
Before this change the request_memoize decorator was not correctly
cleared at the start of a request, because the hook used for this was
unregistered with unregister_plugin_hooks.

13034 FIX Network scan for folders did not show all criticality tags

In "Setup > Hosts" the "Network Scan" section of the folder properties
contain a "Set criticality host tags" dropdown. The values of this
dropdown are cached.  This caching was not correctly cleared in previous
versions. Newly added values did not show up in the list.

Change-Id: Id86b5410840324b0d31d688e9d42b2235461e020