13202 FIX <tt>wmi_cpuload</tt>: Handle Windows time counter resets
The check plugin <tt>wmi_cpuload</tt>, which monitors the CPU load of
Windows hosts, uses an internal Windows time counter to compute the
average load over the last 5 and 15 minutes. This counter can be reset
by Windows due to various reasons.

<tt>wmi_cpuload</tt> can now handle such resets. In case a reset
occurs, the computation of the averages is reset as well and started
from scratch based on the new counter.

CMK-8510

Change-Id: I03aa7c48a9ee091f2911e7d69eda7706068d1b7d