Datadog special agent: rework mechanism for fetching events
Up to now, the agent remembered the last point in time when it fetched
events and used this timestamp as a new starting point for fetching new
events. It turned out that this can lead to missing events.

Now, the agent fetches all events younger that a set maximum age.
Furthermore, it memorizes the ids of the last fetched events and only
forwards events it hasn't seen yet.

SUP-7487

Change-Id: I0cfbef1be2ee00a70d87995982d0fb9598ed1c1a