13208 FIX <tt>qmail_stats</tt>: Check for <tt>/var/qmail/bin/qmail-qstat</tt>
If the <tt>qmail-qstat</tt> binary is not in the PATH of the agent,
no <tt>qmail_qstats</tt> section is produced. On some systems, the
binary was present, but not in the PATH. The Linux, FreeBSD and
OpenWRT agents now append <tt>/var/qmail/bin</tt> to the PATH s.t.
<tt>qmail-qstat</tt> is found on such systems.

Closes #394

Change-Id: I3f2d56c72069f309f4c55e0c0d05dea00486eff2