12959 FIX contact group host attribute in REST API
Previously, the contact_group attribute could not be set, because it
lead to an error. This has been fixed and it can now be set. The REST
API documentation has been updated as well.

Please note that, if you're using a statically typed language and a
generated API client, you may have to regenerate your API client library.

Change-Id: Ia679fd1a6ee2641278ed903a591113c3ed84e98b