Recreate package-lock.json to prevent changes during .f12
Also prevent local mirror URLs from being added to the package-lock.json
now. Previously we got different URLs in the "resolved" entries,
depending on which registry was used for creating the lock file. This
will now be streamlined to the public mirror URLs.

I hope that fix the annoying behavior of npm when .f12. If not, we'll
have to take another look.

Change-Id: I52620ab4c085fa4a2097508d8ad02d02099e8b10