static parameters: pass configured parameters as such
The function '_get_static_check_entries' previously passed the configured
parameters as the argument that ususally contains the discovered parameters.
Since the configured parameters may contain time specific parameters,
(which the discovered parameter can not) a workaround had been implemented:
The time specfic parameters are "moved out of the way", then the default
parameters are added to an empty parameter set, and then the time
specific parameters are added.
Since we can now pass 'configured_parameters', this workaround is no longer needed.

Change-Id: I857e7eb80004a3de5446bd905f04294d9b7b3367