13306 FIX mk_docker.py: Improve error reporting when executed asynchronously
The agent plugin is expected to transport some errors to the Checkmk server to
make them visible on the service discovery page. E.g. when the docker python is
missing on the target system, this should be displayed to the user.

However, when the plugin is executed in asynchronous mode, the error reporting
of the plugin told the agent to drop the error messages in previous versions.

Change-Id: I7a0f2c7a36f076f4a360c03142da17851fa7ad66