Cleanup WSGI application contexts
Previously we used a DummyApplication for some of our tests which seemed
to work fine, but left out important parts of our WSGI application. For
example the profiling, and request related hook middleware was not in
place. Without the later, the cleanup of request scoped memoized
function calls was not working as intended in unit test context.

This was not visible, since `cmk.gui.hooks.request_memoize` is not widely
used at the moment.

Change-Id: I226825d76e0798b9015e525340e1798cc137a978