Ensure request_memoize cleanup also in unit test cases
Previously the request_memoize cache was only cleaned up during
actual request processing, which is fine for the normal execution
of the GUI.

But in many unit tests we don't really handle requests, but only
construct an application and request context to execute the test.

To have a cleanup also in this situation, we now also trigger the
cleanup when exiting our `RequestContext`.

Change-Id: I86d5e5daae0b5cf7347bdd3a37199d4bc07ad432