13277 FIX Lock checkmk configuration while configuration snapshot is unpacked
Background processes like the <tt>Automatic service discovery</tt> and the DCD may automatically
trigger a rebuild of the monitoring core configuration. Due to a race condition, this can also
happen on remote sites while a snapshot is being unpacked, resulting in invalid/incomplete configurations.

It is no longer possible to create the core configuration while a snapshot unpack is in progress.
Instead, the process in question waits until the unpacking has finished.

Change-Id: I6a4db29a38fc4244ad289060e612f8e09de2aaf4