13347 FIX <tt>check_sftp</tt>: Fix command line argument for private key
The active check <tt>check_sftp</tt> claimed to accept <tt>--private-key</tt>
as a command line argument for specifying a private key to be used for the
SFTP login. However, attempting to use this option resulted in the error
<tt>option --private-key not recognized</tt>.

As of this werk, <tt>check_sftp</tt> accepts <tt>--private-key-file</tt> as
a command line argument. This can be used to specify the path to a private
key or certificate file which will be used for authentication.

Closes #398

Change-Id: I8c4f70d2f0831ab4f054283f95d87afdab3def80