13047 FIX RESTAPI: fix agent information and agent download
Before this Werk, both endpoints for downloading the agent binary and
for getting the configuration of a agent shared the same URL, so
effectively only the "show agent information" endpoint to show the
configuration was active.

This endpoint did not work properly and returned an error:
  AttributeError: 'str' object has no attribute 'items'

With this Werk a second endpoint was introduced to download the agent
binary and the endpoint to show the agent configuration was fixed.

SUP-7581

Change-Id: Iafce580bcc8d9ec2968aac83840f1d6a2f828ae7