Update kaspersky_av_client.vbs
The script used a hardcoded "." as date separator. As a result, the script (CDate) failed on non-German locales. This is a second attempt based on the feedback recieved in https://github.com/tribe29/checkmk/pull/385. The script now works with all locales/regional settings because it reads the system's date/time settings from the registry.

A brief explanation:
The CDate function fails if it doesn't recognize the date/time format of the input. This means that input must be formatted the same way as the system's date/time settings, which is different all across the world. There's variation in  date/time separators, order, leading zeros, and 12/24hrs settings. The date/time format of the Kaspersky registry values is always identical: dd-MM-yyyy HH-mm-ss.

To be able to use CDate, we have to convert the Kaspersky date/time values. To achieve this, the script puts both the Kaspersky values and the date/time format from the registry in arrays. The order in the array with dateSTAMP elements is known:  (0) is days, (1) is month, and (2) is the year. By looping through the array with the date FORMAT elements (from the registry), the matching stamps can be stored in a new array in the same order as the date format. Now we can construct a date/time string with valid input to do CDate-based calculations. 

Finally, the script performs one more date/time conversion: from the system's format to the 'German' format checkmk server requires.

The script has been succesfully tested in a prodution environment with 'Dutch' settings. We also ran several test cases with
sDate, sTime, sShortDate, and sTimeFormat settings in HKEY_CURRENT_USER\Control Panel\International\. (including -, /, and . as date separator, shuffeled date order, single/double characters to test leading 0, and both 12/24hr settings). Still, additional testing would be appreciated.