13049 RESTAPI: require accept header
In previous version the accept header of the HTTP-request was completely
ignored.

With this werk the accept header is obligatory. If no or a wrong accept
header is provided an error with a 406 status code will be returned.

For most requests this means you have to provide the follwing header
line:
    Accept: application/json

For some other requests different accept headers may be required.
All headers are documented in the official rest-api documentation.

For example when downloading the agent binary the following header is
required:
    Accept: application/octet-stream

CMK-8660

Change-Id: I8e46b349f3857726a8f8188c52bc254d61456c92