Move useful function to their own file
Currently, the `AttributeListAsIntColumn::createFilter()` implementation
is shared with `AttributeBitmaskLambdaColumn` over inheritance.

This patch moves this implementation out of `AttributeListAsIntColumn`
to break the dependency and remove the artificial coupling and complexity.

The severance of the inheritance relationship is done elsewhere.

Also note that `AttributeListColumn` also uses the filter from
`AttributeListAsIntColumn`.

CMK-7076

Change-Id: I374a9809b977e6d407dfcca309a66c76c7619c05