13128 FIX agent_vsphere: Fix bug occuring when vSphere secret has a special format
When specifying a vSphere secret starting with a dash (a "-"), the vsphere
special agent will run into an error, because the secret will get recognized
as an additional command line argument by it's argument parser.

This is fixed now by passing this argument in a more robust way.

SUP-7465, SUP-7961

Change-Id: I92fe1033dd51ee2507d7ed8fbe691074965e2341