Port AttributeListColumn to new API
In the new column API, `serialize` ist stateless and transforms
each elements of type U (from ListColumn::Callback<T, U>) to string,
this is performed here by mapping the bit index to the string.

Also, in order to use the IntFilter on the `vector<string>` column, we
need to convert the strings back to the bitmask value.

CMK-7076

Change-Id: I7d50d773ada2b5083a99a38bb1c147652686b78d