13077 FIX REST API ETag configuration option
The configuration option for the ETags, previously introduced in
Werk 11938 did not show up in the global configuration setup GUI.

This was due to a bug, which this Werk fixes. The ETag functionality can now
be turned on/off through the "Global settings > Site Management > REST API:
Use HTTP ETags for optimistic locking" configuration value.

Change-Id: I3b7aadd38ac357554f68572ea5927b132558e3ce