Update check_mk_agent.linux
ntpsec is a fork of ntpd which is found in the standard package repositories of a number of distros, including debian and ubuntu.  It behaves in a similar fashion to the ntpd package, replacing timesyncd, and shipping with a ntpq which gives output in the same format.

See https://ntpsec.org/ for more information.

I've been using ntpsec with check_mk_agent under 1.6 with no problems and have used this change to make it work with 2.0 again.