13168 FIX mk_postgres: fix for CentOS 8: "MAIN": not running
The path for postgres instance "main" is called "data" and not "main" on some platforms.
That could lead to the following error message:
instance "MAIN": not running

In order to support more platforms the instance detection for "main" will additionally fallback
to find "data" in the path if "main" was not found.

The agent must be redeployed in order to apply this bugfix.

Change-Id: Ic231258b540df49911e6b86564ccfde8d52890c2