12997 FIX Report scheduler does not remove elements anymore
Reports generated by the scheduler would consume the configured report
elements leading to empty reports if the scheduler is triggered in
sequence, because they would be on the same Apache worker process.
Reappearance of the elements took place if other worker had a clean copy
of the reports to generate.

This is now fixed.

CMK-8266
SUP-7013
SUP-7647
SUP-7737
SUP-7791

Change-Id: I0f366258e6bde46065983ddff57bbca394b78b11