Also prevent incompatibilities with older remote sites
In distributed setups the sync from a newer major version central site
to an older central site is not allowed. Since 2.1.0 the remote site is
performing most of the validations. But in case the newer site is the
central site and the remote site is the older, e.g.
2.1.0 to 2.0.0, there is no validation logic on the remote site. To
ensure the validation is also performed in this case, we execute the
validation logic also in the central site when receiving the answer to
the remote call before processing it's content.

Change-Id: Icb3f736138752e71a484275348ab30ac0a7e8bb2