13360 FIX No longer allow semicolon in service descriptions
When having a semicolon in service description the metric
system would break i.e. metrics for this service would not be displayed.

Having a semicolon in service description was not allowed before 2.0 but
this restriction has been removed with Werk #12799.
From now on the semicolon will again be removed from service descriptions.

With this change all services containing a semicolon will be renamed after
the next "Activate changes"

Notice: You are only affected by this issue if you are using the Checkmk Microcore

CMK-8777

Change-Id: If008cdd4823bce4e29c61729c88bf5f2a5cb95f4