Add to trusted CAs: Write new CA as string to config
There was a mixup of bytes and str for the underlying config variable
trusted_certificate_authorities:

* When editing this global setting directly, the certificates where
stored as str.

* When pressing "Add to trusted CAs" in the livestatus encryption page,
the new certificate was stored as bytes.

In the latter case, there is a check for duplicates to avoid having the
same certificate multiple times in the trust list. This check was also
confused by this bug.

Change-Id: Id1234df62a3523e2b93c9070ad74f39d6e4e3d70