13387 Improve usability of "Create rule" buttons
So far the "Create rule" buttons were positioned at the bottom of a ruleset page and did not fit in with the standard
Checkmk buttons. These buttons are now located at the top of the page as page menu suggestions and yield several more
intuitive options for preselecting folder, host and service/device before entering the rule properties dialog. Also,
"Create rule" is reworded to "Add rule".
(CMK-8518)

Change-Id: I995dc03362f4d392b063db755999c3edcc940ee4