Revert "Revert "12699 Activate Livestatus TCP socket with systemd on recent enough systems""
Fix broken substitution in perl:

`\1` for the RHS substitute does not scare sedders *but* in our case,
the substitute is such as 127.0.0.1, resulting in the RHS `\1127.0.0.1`
and catastrophe.  The more perlish $1, however, can be disambiguated as
`${1}127.0.0.1`: aka this fix.

Also, I like `|` better than `@` as an alternative separatator because
`$@` has a very common meaning in a shell script.

CMK-8683

This reverts commit 589c35a5d49b78333b694f1b46dbc46306fc3cf3.

Change-Id: Ia198a380fd814258c77979a64e5ce35343fec4c4