LIVESTATUS_TCP_* substitute with ed instead of perl
`perl -i` and `sed -i` are not true in place in that the original file
gets unlinked and replaced.  This is bad because the local configuration
under `.config/systemd/user/` is hard linked to the template in
`/etc/mk-livestatus`.  And the hard linking is necessary because systemd
tends to remove soft links.

`ed(1)`, however, makes true in place edits that leave the inode
unchanged.  This is the only way to make the hard linking work as
expected.  An alternative could be `sponge(1)` that could let us keep
perl but it *does not guarantee* that there will never be a copy.

Anyway, the replacements are simple-enough that `ed(1)` is fine.

CMK-8683

Change-Id: I95a0bf20278410e69475f46b2d05e4ad0c20c40b