Revert "Revert "Extend Pipfile tests""
This reverts commit c4f9517ef986a8e334f32f2c3e247c9e8de9fab1 and fixes
the reverted commit (eb18049ca0cd3a83ce98c91a48068f0b3d45c14d).

We want to unpin the dependencies in master and only pin in the stable
branches. Therefore we can additionally remove sub-dependencies. To
check that we are not missing dependencies a test was added to check for
"undeclared" dependencies.
To check for unneeded dependencies another test was created.

CMK-8716

Change-Id: I06d18a8b87838c45064fa63d4343f8f7dfe765db