13361 FIX CRE: Disallow round brackets and tab in service descriptions
When having a round bracket in service description the nagios core would
not start.

Having a round bracket in service description was not allowed before 2.0.0p14 but
this restriction was accidentally removed with Werk #13360.
Round brackets will again be removed from service descriptions from now on which
matches the behaviour from 2.0.0p13.

Notice: You are only affected by this issue if you are using the Nagios core

Change-Id: Ia575f6f808f2af3ad9773961ab9bb70bfc390744