Simplify plugin loading a bit
Previously the GUI was loading all the GUI plugins during the first
request made in an apache process. In case this was a AJAX graph update
(ajax_graph.py) the loading of the plugins could dramatically slow down
the request and result in a bad user experience of graph updates.

Since we are about to move the plugin loading to the application import
time, this optimization can easily be dropped.

Change-Id: If5d067dc5ed25e78cd8b04a7dbc0abdd9e74cd3f