Bypass creation of AutocheckService instances
The service classes currently are a mess, because they
sometimes contain the fully computed parameters, and
sometimes only the discovered parameters.

Since in this case we don't need either of them, it's best
to just load the autocheck entries, and compute the
description separately.

Note that we now assume that we can in fact compute the
service description of all affected autochecks, and do not
silently skip them upon an exception raised by
'config.service_description'.

Change-Id: I225d06de163c366a339ca85edd3d121ad6ab9350