13080 FIX REST API: allow setting of downtime from read-only site
When having a read-only site attached to the master site, previously
it wasn't possible to set downtimes from there, even though it should
have been possible.

This has been changed to allow setting of downtimes even if the host is
not or no longer configured. It only needs to be currently monitored.

Change-Id: I41b25b405f0d9288b210d39451002d98332d4a21