Fixed network <-> host byte order confusion for ports.
Things are quite simple now: The conversion from host byte order to network
byte order is done exclusively at the place where it is supposed to be done,
i.e. while serialising data for a syscall.  In all other places it is just
a a plain old integer with the value one expects.

Change-Id: I4e783ae0ead61c5b4e99fd121c45b5d209a72609