Improved port typing a tiny bit.
We now consistently use uint16_t when a port number is in host byte order
(i.e. a plain old number), and in_port_t only when it is in network byte
order (i.e. never, at least in our own code).

Thanks to ancient APIs, the compiler is not of much help here, both are
eventually typedef'd to the same underlying integer type.  At least
readability is improved a bit.

Change-Id: I8259d32a39245f5a526fb9a18caeeb4b07e4718d