Separate GUI frontend .f12 from backend
Previously a .f12 in cmk/gui also triggered an update of the web files.
While this might be useful when working in both places simultaneously,
it is annoying when only working on the server side python code.

Both are now again separated. For the developers we now have the
following tools at hand:

* When working with cmk/gui, you now have faster deploys with the local .f12.
* When working on the frontend code, use scripts/run-webpack-watch.
* When working on other stuff in web (e.g. images), use web/.f12

Change-Id: I66fd5be158be604637a9c612bc8bf2446d62fc30