Fix test_plugin_vs_wato unit test
The test did not work as intended in all situations because it relied on
GUI plugins being loaded which normally done by GUI specific fixtures
(which are defined in tests/unit/cmk/gui). Because of the previous
location of the test, the environment was not set-up correctly.

By simply relocating the test this specific issue could be solved.

Another issue occured with this combination of tests:

```
pytest -xsvv -T unit \
    tests/unit/checks/test_generic_legacy_conversion.py \
    tests/unit/cmk/gui/plugins/wato/check_parameters/test_plugin_vs_wato.py
```

This change also fixes the new findings uncovered by this cleanup.

Change-Id: Ica87cb8572a7edf2da82158e3fa62266fa1bd7fa