13518 FIX Liveproxy Daemon: React to changes in CAs without manual restart
When using encrypted connections, the CAs trusted by the Liveproxy Daemon
(LPD) can be administrated via the global setting "Trusted certificate
authorities for SSL". Depending on the current state of the LPD, a manual
restart was necessary for the LPD to take changes in this setting into
account. In particular, removing a CA of a remote side to which the LPD
was currently connected to did not result in a loss of connection without
a restart.

CMK-8933

Change-Id: Ieb3025b64ee4474625d453afa1c3692c138d9e7d