13473 FIX Do not drop services of incompatibly changed plugins
AKA: Don't raise in config.service_description

This changes the behaviour if an unexpected formatter is
found in an service description template. We will no longer
raise an exception.
Judging from the exception message, this was meant to deal
with the case of a service having been discovered at a time
when the plugin did not have an item, and then being read
after the plugin has been changed to have an item.

In this (rare) case we will just render "<missing an item>"
as fallback.

A more reliable service_description function will allow
for various cleanups, as we no longer have to precompute
the description, and pass it around with the service.
(The attribute and the service_description function are
then used inconsistently.)

Note that a service description depends not only on the
service, but also on the configuration (and thus the host),
hence we prefer the function approach.

Change-Id: Ief7b37e0d663356ef4ce5eb2b7758128bfdcc1cc